relocate scss files into /styles, rename tsx pages
