+++ src/components/CommentCompose.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { IonAvatar, IonButton, IonCard, IonCardContent, IonIcon, IonTextarea } from '@ionic/react';
-import './CommentCompose.scss';
+import '../styles/components/CommentCompose.scss';
 import { IComment } from '../services/models/DocTypes.model';
 import { db, FieldValue, Timestamp } from '../services/data/firebase';
 import { arrowForward } from 'ionicons/icons';
