+++ src/pages/ProfileView.tsx
@@ -9,7 +9,7 @@ import { db } from '../services/data/firebase';
 import Post from '../services/FirebasePost.service';
 import PrimaryUser from '../services/PrimaryUser.service';
 import User from '../services/User.service';
-import './ProfileView.scss';
+import '../styles/pages/ProfileView.scss';
 
 const ProfileView: React.FC = () => {
     const [user, setUser] = useState<User | PrimaryUser | undefined>();
