+++ src/styles/components/HomePostView.scss
