+++ src/styles/components/LandingInfo.scss
