+++ src/styles/components/PostContainer.scss
