+++ src/styles/pages/ProfileView.scss
